+++ src/components/__generated__/EditFieldsDialog_field.ts
@@ -1,0 +1,20 @@
+/* tslint:disable */
+/* eslint-disable */
+// @generated
+// This file was automatically generated and should not be edited.
+
+// ====================================================
+// GraphQL fragment: EditFieldsDialog_field
+// ====================================================
+
+export interface EditFieldsDialog_field {
+  __typename: "Field";
+  /**
+   * The ID of an object
+   */
+  id: string;
+  /**
+   * Name of the field
+   */
+  name: string;
+}