+++ src/pages/__generated__/ModelQuery.ts
@@ -118,7 +118,7 @@ export interface ModelQuery_model {
   /**
    * Name of this card model (e.g. "Basic", "Basic with Reversed")
    */
-  name: string | null;
+  name: string;
   /**
    * Fields associated with this model
    */